Addition of the requirements.txt
In this commit I created the requirements of the code based on the imports of the network. First the notebooks are going to be tested and then we will transfer the code necessities into the requirements.txt

I have tried to create a CONDA environment with the requirements but it was not possible due to the fact that there is not enough space. Also it was tried to create a python venv but the same error occured. I contacted with Josep Pujal.

puta vida ni un environment se fer.