# Comments on functions
The System class is the class that has all of the information of the combination of the model and prompts. We are putting some comments in this class to understand what are all of its components.

Now we asked Github Copilot to put some comments about the hooks, which is something that it is not very easy to understand