Bug/fix token length (#9)
* process the transcriptions

What we need to do:
Crop or pad the transcriptions to have all the same length.

Why?
We need all the transcriptions to be the same size to be processed by PyTorch data loaders

* For now, we do zero padding but the segments of both audio and text are random...

* Now we have a problem with DataParellalization, it makes the model not callable. We will have to fix this and ask help of our master, Federico Costa.

* some changes in all of the files

* there was an idention bug in the train single epoch

* # Shape problem

We are encountering a shape problem. Expected 1024, 80,3 but I am inputing 1,32,480000 This is because I am inputing 1, batch_size, sampling_rate*seconds

* we will do mel spectrograms processing. it is required

* Still fixing the issue with the spectrogram creation

* In theory it shouldn't give a shape problem

* Solved the issue with shape, now we have a problem with the data type, it shold be both Long or Int and it is CUDA Float

* last pushes

* # Long Tensor bug solved.

I was passing to Whisper the tensors like this whisper(input, softprompt) and they need to be concatenated into one single vector like whisper(concat(input,softprompt)).

Now this error doesn't exist anymore but I have another error in which Whisper.forward() missing 1 required positional argument: 'tokens'

* # Whisper decoder input

I am trying to implementate the whisper decoder input. For doing so I will need a lot of beer.

* # Trying to solve the problem with the dataset and whisper softprompting

As far as I know, the whisper decoder needs some inputs that will be fed using the dataset class.

* lucas is recoveredgit add .!