# Problem with shapes in the __getitem__
Now we have a problem with the dataset bc we have different shapes in the same batch and we cannot use data loaders. The part that is having this problem is the one related to the text. I suppose that this is due to the transcription, so we will have to normalize.

Also it is interesting to see the strategies used by the other.