now I have everything padded :)
The next problem is that I am predicting the tokens and this is not actually correct! In theory we should be predicting the probabilities, idk how this works in ASR, I don't think we are predicting one-hot encoders all the time, so I assume that we will have to do something different! Maybe I am wrong and we are predicting one-hot tensors all the time idk...

we will have to review the theory...